bin/requirement.txt
asn1crypto==0.24.0
biopython==1.79
cryptography==2.1.4
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
numpy==1.19.5
pycrypto==2.6.1
pygobject==3.26.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
